requirements.txt
asgiref==3.9.1
Django==5.2.5
django-unfold==0.65.0
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.15.2